requirements-dev.txt
pytest>=8.0.0
pytest-cov>=5.0.0
moto[eks,rds,elasticache,sqs,s3,ec2]>=5.0.0
black>=24.0.0
ruff>=0.4.0
mypy>=1.10.0
boto3-stubs[eks,rds,elasticache,sqs,s3,ec2,ce]>=1.34.0
